// src/store/adminNotificationStore/action.enum.ts
enum actionTypes {
  GET_ADMIN_NOTIFICATIONS_SUCCESS = "GET_ADMIN_NOTIFICATIONS_SUCCESS",
  GET_ADMIN_NOTIFICATIONS_PENDING = "GET_ADMIN_NOTIFICATIONS_PENDING",
  GET_ADMIN_NOTIFICATIONS_FAILED = "GET_ADMIN_NOTIFICATIONS_FAILED",

  GET_HEADER_ADMIN_NOTIFICATIONS_SUCCESS = "GET_HEADER_ADMIN_NOTIFICATIONS_SUCCESS",
  GET_HEADER_ADMIN_NOTIFICATIONS_PENDING = "GET_HEADER_ADMIN_NOTIFICATIONS_PENDING",
  GET_HEADER_ADMIN_NOTIFICATIONS_FAILED = "GET_HEADER_ADMIN_NOTIFICATIONS_FAILED",

  READ_ADMIN_NOTIFICATIONS_SUCCESS = "READ_ADMIN_NOTIFICATIONS_SUCCESS",
  READ_ADMIN_NOTIFICATIONS_PENDING = "READ_ADMIN_NOTIFICATIONS_PENDING",
  READ_ADMIN_NOTIFICATIONS_FAILED = "READ_ADMIN_NOTIFICATIONS_FAILED",

  DELETE_ADMIN_NOTIFICATIONS_SUCCESS = "DELETE_ADMIN_NOTIFICATIONS_SUCCESS",
  DELETE_ADMIN_NOTIFICATIONS_PENDING = "DELETE_ADMIN_NOTIFICATIONS_PENDING",
  DELETE_ADMIN_NOTIFICATIONS_FAILED = "DELETE_ADMIN_NOTIFICATIONS_FAILED",
}

export default actionTypes;